requirements.txt
discord.py==2.3.2
libretranslatepy==2.1.3
lingua-language-detector==2.0.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
